fix: restructuring files/directories
