+++ src/pages/subpage/Skills.jsx
@@ -1,0 +1,1 @@
+export function Skills() {}
